[refactor] Move crypto functions to separated folder
