+++ src/crypto/ecc.c
