+++ src/crypto/prng.c
